chore: utils moved to separate folder
